Add files via upload
The executable file is from MATLAB Runtime which generates an HDF5 file with  wrapped and unwrapped phase screens as well as the location of brightest intensity with that output signal. function [A,B,C] = Data_Gen(Rytov,wvl,Lout,Dz,N,deltan,nscreen,ntrial)

The python script is capable of calling the executable and splitting the matrices into a list of matrices. However I am not fully understand subprocess and what I want to do is ask the user the input values, store then and use them in my subprocess rather than having to manually change it on the script.