Add files via upload
Open a MATLAB R2017a Runtime executable file that does simulated laser propagation through the atmosphere in the Kolmogorov spectrum. It will output a single HDF5 file that contains unwrapped and wrapped phase matrices which we save as a png file. The python code will iterate through the executable file 1000 times. at the end of it, it'll move all the HDF5 files into a specified directory for later use.